Use full pyramid function by default
Saving incremental pyramd in CutPyramid class for possible use in later debugging
